+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning].
 - exclude and extend-exclude for excluding files and directories
 - add a subcommand hooks, to be used with e.g. pre-commit
 - add a hook check-large-files to check for large files after cleaning notebooks
+- add `--isolated` flag to ignore configuration files
 
 ## [0.3.7] - 2024-06-11